update warnings and filedialog

+++ libs/bcg_opengl/events/bcg_events_spatial_index.h
@@ -16,7 +16,7 @@ struct setup_kdtree{
 
 struct setup_octree{
     entt::entity id;
-    int leaf_size = 10;
+    size_t leaf_size = 10;
     int max_depth = 21;
 };
 
@@ -24,7 +24,7 @@ struct setup_sampling_octree{
     entt::entity id;
     sampling_octree::SamplingType type;
     std::vector<size_t> indices;
-    int leaf_size = 10;
+    size_t leaf_size = 10;
     int max_depth = 21;
 };
